Strix 2.5.0
Feats
1. 优化应用初始化时SQL查询数量.
2. 统一使用Lambda语法进行数据库查询.
3. 优化代码规范.

Fixes
1. 修复WebConfig未生效的问题.

Deps

+++ src/main/java/cn/projectan/strix/controller/system/SystemManagerController.java
@@ -3,7 +3,6 @@ package cn.projectan.strix.controller.system;
 import cn.projectan.strix.controller.system.base.BaseSystemController;
 import cn.projectan.strix.core.cache.SystemMenuCache;
 import cn.projectan.strix.core.cache.SystemPermissionCache;
-import cn.projectan.strix.core.cache.SystemRegionCache;
 import cn.projectan.strix.core.ret.RetBuilder;
 import cn.projectan.strix.core.ret.RetResult;
 import cn.projectan.strix.core.validation.group.InsertGroup;
@@ -24,23 +23,26 @@ import cn.projectan.strix.model.response.system.manager.SystemManagerResp;
 import cn.projectan.strix.service.SystemManagerRoleService;
 import cn.projectan.strix.service.SystemManagerService;
 import cn.projectan.strix.utils.*;
-import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import com.baomidou.mybatisplus.core.conditions.update.LambdaUpdateWrapper;
 import com.baomidou.mybatisplus.core.conditions.update.UpdateWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.util.Assert;
+import org.springframework.util.CollectionUtils;
 import org.springframework.util.StringUtils;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.stream.Collectors;
 
 /**
+ * 系统人员
+ *
  * @author ProjectAn
  * @date 2021/6/11 17:40
  */
@@ -54,51 +56,49 @@ public class SystemManagerController extends BaseSystemController {
     private final SystemManagerRoleService systemManagerRoleService;
     private final SystemMenuCache systemMenuCache;
     private final SystemPermissionCache systemPermissionCache;
-    private final SystemRegionCache systemRegionCache;
     private final RedisUtil redisUtil;
 
+    /**
+     * 查询人员列表
+     */
     @GetMapping("")
     @PreAuthorize("@ss.hasPermission('system:manager')")
     @StrixLog(operationGroup = "系统人员", operationName = "查询人员列表")
     public RetResult<SystemManagerListResp> getSystemManagerList(SystemManagerListReq req) {
-        QueryWrapper<SystemManager> systemManagerQueryWrapper = new QueryWrapper<>();
-        RegionPermissionTool.appendRegionPermissionToQueryWrapper("region_id", systemManagerQueryWrapper);
-        if (StringUtils.hasText(req.getKeyword())) {
-            systemManagerQueryWrapper.like("nickname", req.getKeyword())
-                    .or(q -> q.like("login_name", req.getKeyword()));
-        }
-        if (NumUtil.checkCategory(req.getStatus(), NumCategory.NON_NEGATIVE)) {
-            systemManagerQueryWrapper.eq("status", req.getStatus());
-        }
-        if (NumUtil.checkCategory(req.getType(), NumCategory.POSITIVE)) {
-            systemManagerQueryWrapper.eq("type", req.getType());
-        }
-        systemManagerQueryWrapper.orderByAsc("create_time");
-
-        Page<SystemManager> page = systemManagerService.page(req.getPage(), systemManagerQueryWrapper);
+        List<String> loginManagerRegionPermissions = loginManagerRegionPermissions();
+        Page<SystemManager> page = systemManagerService.lambdaQuery()
+                .eq(StringUtils.hasText(req.getKeyword()), SystemManager::getNickname, req.getKeyword())
+                .or(StringUtils.hasText(req.getKeyword()), q -> q.like(SystemManager::getLoginName, req.getKeyword()))
+                .eq(NumUtil.checkCategory(req.getStatus(), NumCategory.NON_NEGATIVE), SystemManager::getStatus, req.getStatus())
+                .eq(NumUtil.checkCategory(req.getType(), NumCategory.POSITIVE), SystemManager::getType, req.getType())
+                .in(!CollectionUtils.isEmpty(loginManagerRegionPermissions), SystemManager::getRegionId, loginManagerRegionPermissions)
+                .orderByAsc(SystemManager::getCreateTime)
+                .page(req.getPage());
 
         SystemManagerListResp resp = new SystemManagerListResp(page.getRecords(), page.getTotal());
 
         return RetBuilder.success(resp);
     }
 
+    /**
+     * 查询人员信息
+     */
     @GetMapping("{managerId}")
     @PreAuthorize("@ss.hasPermission('system:manager')")
     @StrixLog(operationGroup = "系统人员", operationName = "查询人员信息")
     public RetResult<SystemManagerResp> getSystemManager(@PathVariable String managerId) {
-        Assert.notNull(managerId, "参数错误");
         SystemManager systemManager = systemManagerService.getById(managerId);
         Assert.notNull(systemManager, "系统人员信息不存在");
-        RegionPermissionTool.check(systemManager.getRegionId());
+        checkLoginManagerRegionPermission(systemManager.getRegionId());
 
-        QueryWrapper<SystemManagerRole> systemManagerRoleQueryWrapper = new QueryWrapper<>();
-        systemManagerRoleQueryWrapper.select("system_role_id");
-        systemManagerRoleQueryWrapper.eq("system_manager_id", managerId);
-        List<String> systemManagerRoleIds = systemManagerRoleService.listObjs(systemManagerRoleQueryWrapper, Object::toString);
+        List<String> systemManagerRoleIds = systemManagerService.getRoleIdListByManagerId(managerId);
 
         return RetBuilder.success(new SystemManagerResp(systemManager.getId(), systemManager.getNickname(), systemManager.getLoginName(), systemManager.getStatus(), systemManager.getType(), systemManager.getRegionId(), systemManager.getCreateTime(), String.join(",", systemManagerRoleIds)));
     }
 
+    /**
+     * 更改人员信息
+     */
     @PostMapping("modify/{managerId}")
     @PreAuthorize("@ss.hasPermission('system:manager:update')")
     @StrixLog(operationGroup = "系统人员", operationName = "更改人员信息", operationType = SysLogOperType.UPDATE)
@@ -107,47 +107,40 @@ public class SystemManagerController extends BaseSystemController {
         SystemManager systemManager = systemManagerService.getById(managerId);
         Assert.notNull(systemManager, "系统人员信息不存在");
         Assert.isTrue(BuiltinConstant.NO == systemManager.getBuiltin(), "内置用户不允许修改");
-        RegionPermissionTool.check(systemManager.getRegionId());
+        checkLoginManagerRegionPermission(systemManager.getRegionId());
 
-        UpdateWrapper<SystemManager> systemManagerUpdateWrapper = new UpdateWrapper<>();
-        systemManagerUpdateWrapper.eq("id", managerId);
+        LambdaUpdateWrapper<SystemManager> systemManagerUpdateWrapper = new LambdaUpdateWrapper<>();
+        systemManagerUpdateWrapper.eq(SystemManager::getId, managerId);
 
         AtomicBoolean needReturnNewData = new AtomicBoolean(false);
 
         switch (req.getField()) {
             case "status" -> {
                 Assert.isTrue(SystemManagerStatus.valid(Integer.valueOf(req.getValue())), "参数错误");
-                systemManagerUpdateWrapper.set("status", req.getValue());
+                systemManagerUpdateWrapper.set(SystemManager::getStatus, req.getValue());
                 Assert.isTrue(systemManagerService.update(systemManagerUpdateWrapper), "修改失败");
             }
             case "type" -> {
                 Assert.isTrue(SystemManagerType.valid(Integer.parseInt(req.getValue())), "参数错误");
-                systemManagerUpdateWrapper.set("type", req.getValue());
+                systemManagerUpdateWrapper.set(SystemManager::getType, req.getValue());
                 Assert.isTrue(systemManagerService.update(systemManagerUpdateWrapper), "修改失败");
             }
             case "role" -> {
                 // 修改管理用户的角色
-                QueryWrapper<SystemManagerRole> systemManagerRoleQueryWrapper = new QueryWrapper<>();
-                systemManagerRoleQueryWrapper.select("system_role_id");
-                systemManagerRoleQueryWrapper.eq("system_manager_id", managerId);
-                List<String> systemManagerRoleIds = systemManagerRoleService.listObjs(systemManagerRoleQueryWrapper, Object::toString);
+                List<String> systemManagerRoleIds = systemManagerService.getRoleIdListByManagerId(managerId);
                 KeyDiffUtil.handle(systemManagerRoleIds, Arrays.asList(req.getValue().split(",")),
                         (removeKeys) -> {
-                            QueryWrapper<SystemManagerRole> removeQueryWrapper = new QueryWrapper<>();
-                            removeQueryWrapper.eq("system_manager_id", managerId);
-                            removeQueryWrapper.in("system_role_id", removeKeys);
-                            Assert.isTrue(systemManagerRoleService.remove(removeQueryWrapper), "移除该管理用户的角色失败");
+                            Assert.isTrue(
+                                    systemManagerRoleService.lambdaUpdate()
+                                            .eq(SystemManagerRole::getSystemManagerId, managerId)
+                                            .in(SystemManagerRole::getSystemRoleId, removeKeys)
+                                            .remove(),
+                                    "移除该管理用户的角色失败");
                         },
                         (addKeys) -> {
-                            List<SystemManagerRole> systemManagerRoleList = new ArrayList<>();
-                            addKeys.forEach(k -> {
-                                SystemManagerRole systemManagerRole = new SystemManagerRole();
-                                systemManagerRole.setSystemManagerId(managerId);
-                                systemManagerRole.setSystemRoleId(k);
-                                systemManagerRole.setCreateBy(loginManagerId());
-                                systemManagerRole.setUpdateBy(loginManagerId());
-                                systemManagerRoleList.add(systemManagerRole);
-                            });
+                            List<SystemManagerRole> systemManagerRoleList = addKeys.stream()
+                                    .map(k -> new SystemManagerRole(managerId, k))
+                                    .collect(Collectors.toList());
                             Assert.isTrue(systemManagerRoleService.saveBatch(systemManagerRoleList), "增加该角色的菜单权限失败");
                         },
                         () -> {
@@ -164,23 +157,22 @@ public class SystemManagerController extends BaseSystemController {
         }
 
         if (needReturnNewData.get()) {
-            QueryWrapper<SystemManagerRole> systemManagerRoleQueryWrapper = new QueryWrapper<>();
-            systemManagerRoleQueryWrapper.select("system_role_id");
-            systemManagerRoleQueryWrapper.eq("system_manager_id", managerId);
-            List<String> systemManagerRoleIds = systemManagerRoleService.listObjs(systemManagerRoleQueryWrapper, Object::toString);
-
+            List<String> systemManagerRoleIds = systemManagerService.getRoleIdListByManagerId(managerId);
             return RetBuilder.success(new SystemManagerResp(systemManager.getId(), systemManager.getNickname(), systemManager.getLoginName(), systemManager.getStatus(), systemManager.getType(), systemManager.getRegionId(), systemManager.getCreateTime(), String.join(",", systemManagerRoleIds)));
         }
 
         return RetBuilder.success();
     }
 
+    /**
+     * 新增人员
+     */
     @PostMapping("update")
     @PreAuthorize("@ss.hasPermission('system:manager:add')")
     @StrixLog(operationGroup = "系统人员", operationName = "新增人员", operationType = SysLogOperType.ADD)
     public RetResult<Object> update(@RequestBody @Validated(InsertGroup.class) SystemManagerUpdateReq req) {
         Assert.notNull(req, "参数错误");
-        RegionPermissionTool.check(req.getRegionId());
+        checkLoginManagerRegionPermission(req.getRegionId());
 
         SystemManager systemManager = new SystemManager(
                 req.getNickname(),
@@ -199,16 +191,18 @@ public class SystemManagerController extends BaseSystemController {
         return RetBuilder.success();
     }
 
+    /**
+     * 修改人员
+     */
     @PostMapping("update/{managerId}")
     @PreAuthorize("@ss.hasPermission('system:manager:update')")
     @StrixLog(operationGroup = "系统人员", operationName = "修改人员", operationType = SysLogOperType.UPDATE)
     public RetResult<Object> update(@PathVariable String managerId, @RequestBody @Validated(UpdateGroup.class) SystemManagerUpdateReq req) {
-        Assert.hasText(managerId, "参数错误");
         Assert.notNull(req, "参数错误");
         SystemManager systemManager = systemManagerService.getById(managerId);
         Assert.notNull(systemManager, "系统人员信息不存在");
         Assert.isTrue(BuiltinConstant.NO == systemManager.getBuiltin(), "内置用户不允许修改");
-        RegionPermissionTool.check(systemManager.getRegionId());
+        checkLoginManagerRegionPermission(systemManager.getRegionId());
 
         UpdateWrapper<SystemManager> updateWrapper = UpdateConditionBuilder.build(systemManager, req);
         UniqueDetectionTool.check(systemManager);
@@ -219,22 +213,24 @@ public class SystemManagerController extends BaseSystemController {
         return RetBuilder.success();
     }
 
+    /**
+     * 删除人员
+     */
     @PostMapping("remove/{managerId}")
     @PreAuthorize("@ss.hasPermission('system:manager:remove')")
     @StrixLog(operationGroup = "系统人员", operationName = "删除人员", operationType = SysLogOperType.DELETE)
     public RetResult<Object> remove(@PathVariable String managerId) {
-        Assert.hasText(managerId, "参数错误");
         SystemManager systemManager = systemManagerService.getById(managerId);
         Assert.notNull(systemManager, "系统人员信息不存在");
         Assert.isTrue(BuiltinConstant.NO == systemManager.getBuiltin(), "内置用户不允许修改");
-        RegionPermissionTool.check(systemManager.getRegionId());
+        checkLoginManagerRegionPermission(systemManager.getRegionId());
 
         systemManagerService.removeById(systemManager);
 
         // 删除管理人员和角色间关系
-        QueryWrapper<SystemManagerRole> deleteManagerRoleRelationQueryWrapper = new QueryWrapper<>();
-        deleteManagerRoleRelationQueryWrapper.eq("system_manager_id", systemManager.getId());
-        systemManagerRoleService.remove(deleteManagerRoleRelationQueryWrapper);
+        systemManagerRoleService.lambdaUpdate()
+                .eq(SystemManagerRole::getSystemManagerId, systemManager.getId())
+                .remove();
 
         // 使登录Token失效
         Object existToken = redisUtil.get("strix:system:manager:login_token:login:id_" + systemManager.getId());
